Increase hash size (#481)
Bench: 5275334

STC

ELO   | 0.85 +- 1.51 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=8MB
LLR   | -0.86 (-2.94, 2.94) [0.00, 3.00]
GAMES | N: 96072 W: 22798 L: 22564 D: 50710

LTC

ELO   | -1.33 +- 3.64 (95%)
SPRT  | 40.0+0.40s Threads=1 Hash=64MB
LLR   | -1.43 (-2.94, 2.94) [0.00, 3.00]
GAMES | N: 15624 W: 3466 L: 3526 D: 8632

SMP

ELO   | 1.91 +- 1.52 (95%)
SPRT  | 4.0+0.04s Threads=8 Hash=64MB
LLR   | 2.95 (-2.94, 2.94) [0.00, 3.00]
GAMES | N: 93713 W: 22250 L: 21736 D: 49727

+++ src/transposition.c
@@ -99,7 +99,7 @@ inline void TTPrefetch(uint64_t hash) {
 
 inline TTEntry* TTProbe(uint64_t hash, int* hit) {
   TTEntry* bucket    = TT.buckets[TTIdx(hash)].entries;
-  uint16_t shortHash = (uint16_t) hash;
+  uint32_t shortHash = (uint32_t) hash;
 
   for (int i = 0; i < BUCKET_SIZE; i++) {
     if (bucket[i].hash == shortHash || !bucket[i].depth) {
@@ -122,7 +122,7 @@ inline TTEntry* TTProbe(uint64_t hash, int* hit) {
 
 inline void
 TTPut(TTEntry* tt, uint64_t hash, int depth, int16_t score, uint8_t bound, Move move, int ply, int16_t eval, int pv) {
-  uint16_t shortHash = (uint16_t) hash;
+  uint32_t shortHash = (uint32_t) hash;
 
   if (score >= TB_WIN_BOUND)
     score += ply;
